Add secure naming negative tests. (#23296)
* Add secure naming negative tests.

* Refactor tests

* Small fix.

* Small fix.

* Small nit fix.

* Small fix.

* Small fix.
